# Minor performance improvement for Point