* Simplify Point `__init__` to speed up instantiation
* Use `__slots__` to shave off a few ns.

Supersedes https://github.com/googlei18n/glyphsLib/pull/450.