@madig 2018-11-16
New benchmark:

Without `__slots__`:
```
In [2]: %timeit Point(100, 200)
364 ns ± 2.03 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [3]: %timeit Point(100, 200)
367 ns ± 6.27 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

With `__slots__`:
```
In [3]: %timeit Point(100, 200)
326 ns ± 0.59 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit Point(100, 200)
329 ns ± 2.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```